+++ docs/changes.md
@@ -5,6 +5,12 @@ title: Parser Change Log
 
 ## Version 3.1 {#v3-1}
 
+### v3.1.2 {#v3-1-2}
+
+* Introduce `Orchestra\Parser\Xml\Document::rebase()` in order to change the parent node for faster parsing.
+* Introduce `Orchestra\Parser\Xml\Document::namespaced()` to fetch namespaced elements.
+* Add support to access namespaced elements via collection by using `item/ns[value1,value2]` etc.
+
 ### v3.1.1 {#v3-1-1}
 
 * Add functionality to swap attribute key with an alias using `item::attribute>attribute`.
@@ -15,6 +21,12 @@ title: Parser Change Log
 * Update support to Laravel Framework v5.1.
 
 ## Version 3.0 {#v3-0}
+
+### v3.0.2 {#v3-0-2}
+
+* Introduce `Orchestra\Parser\Xml\Document::rebase()` in order to change the parent node for faster parsing.
+* Introduce `Orchestra\Parser\Xml\Document::namespaced()` to fetch namespaced elements.
+* Add support to access namespaced elements via collection by using `item/ns[value1,value2]` etc.
 
 ### v3.0.1 {#v3-0-1}